LifeReloaded/requirements.txt
fastapi==0.103.1
jinja2==3.1.2
pydantic==1.10.12
python-dotenv==1.0.0
google-generativeai>=0.3.0
redis>=6.0.0
uvicorn==0.23.2
